og-auto-new/requirements.txt
web3==6.19.0
python-dotenv==1.0.1
eth-account==0.10.0
hexbytes==0.3.1
eth-abi==5.1.0
py-solc-x==2.0.3
requests==2.32.3
pydantic==2.9.1
tenacity==9.0.0
jinja2>=3.1.4
py-solc-x>=2.0.3